Create new page and link to homepage
In this iteration, I created a new folder within the repository for the
different recipe pages that will be linked to the homepage. I also
created the first recipe page (chipotle-oregon-burritos.html) and laid
down the basic structure, which will have recipe content added in the
next iteration. In order to keep up with the new page iteration, I
updated index.html with a link to chipotle-oregon-burritos.html.