Add page content for Burrito recipe
In this iteration, I added an image of the finished dish, a
description, an ingredients list, and steps for the Chipotle Oregon
Burritos recipe. To distinguish the ingredients from the steps, I
used an unordered list for the recipes and an ordered list from the
steps as appropriate given context.