Add protein cup recipe with content, update homepage
In this iteration, I added a new page for a Strawberry Protein Cup
recipe, the contents of which I added in appropriate formatting which
matches the structure I used for the two recipe pages added in prior
commits. To keep the homepage consistent with this iteration, I moved
the recipe links on the homepage into an unordered list and added a link
for the Strawberry Protein Cup recipe.